% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_dist_to_colony}
\alias{add_dist_to_colony}
\title{Add distance to colony.}
\usage{
add_dist_to_colony(dt)
}
\arguments{
\item{dt}{Tracking data as data.table. Should contain columns \code{latitude},
\code{longitude}, \code{colony_latitude}, \code{colony_longitude}.}
}
\value{
Nothing. Adds distance (in meters) in place as column
\code{calc_distance_to_colony}.
}
\description{
Calculates the distance from the GPS position to the colony.
}
\examples{
\dontrun{
add_dist_to_colony(tracking_data)
}
}
