% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_speed}
\alias{add_speed}
\title{Add speed.}
\usage{
add_speed(dt)
}
\arguments{
\item{dt}{Tracking data as data.table. Should contain column
\code{calc_distance_diff} and \code{calc_time_diff}.}
}
\value{
Nothing. Data is added in place as column \code{calc_speed_2d}.
}
\description{
Calculates the average 2 dimensional speed of the individual since the
previous GPS fix.
}
\examples{
\dontrun{
add_speed(tracking_data)
}
}
