% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_time_since_previous_fix}
\alias{add_time_since_previous_fix}
\title{Add time since previous fix}
\usage{
add_time_since_previous_fix(datatable)
}
\arguments{
\item{datatable}{A data.table with tracking data. Should at least include
a column `device_info_serial` and `date_time`}
}
\value{
a new datatable with the time difference column (`calc_time_diff`) added to it.
}
\description{
Calculates the time (in seconds) since the birds last fix.
Data is first ordered by individual and date_time. Next, for each individual
the time difference between a fix and its previous fix is calculated.
}
\examples{
\dontrun{
add_time_since_previous_fix(tracking_data)
}
}

