% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_year_month_hour}
\alias{add_year_month_hour}
\title{Add year, month, hour columns}
\usage{
add_year_month_hour(data)
}
\arguments{
\item{dt}{Tracking data as a data.table.}
}
\value{
Nothing, columns are added in place
}
\description{
Add the year, month and hour of the GPS fix in separate columns
}
\examples{
\dontrun{
add_year_month_hour(tracks_data)
}
}

