% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds_validated}
\alias{birds_validated}
\title{Birds validated}
\format{data table}
\usage{
birds_validated
}
\description{
An example data table containing validated bird metadata (only 2 records).
This is what you can expect after successfully loading and validating bird
metadata.
}
\keyword{datasets}

