% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich_data}
\alias{enrich_data}
\title{Enrich data}
\usage{
enrich_data(tracking_data, bird_data, corine_raster_data)
}
\arguments{
\item{tracking_data}{Data table obtained by using the validate_tracks_data
function}

\item{bird_data}{Bird metadata obtained by using the validate_bird_data
function}

\item{corine_raster_data}{Filename containing corine raster data to be joined
with bird tracking data.}
}
\value{
Data table containing enriched data
}
\description{
Enrich the bird tracking data by precalculating attributes
and joining data with other sources. See the package vignette for a complete
description of the procedure.
}
\examples{
\dontrun {
enrich_data(tracking_data, bird_data)
}
}

