% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbbg_gps}
\alias{lbbg_gps}
\title{Sample of Movebank GPS data of a Lesser black-backed
gulls bird tracking dataset}
\format{
A tibble (data.frame) with 100 rows and 22 variables.
}
\source{
To download the entire file (zip format): \url{https://zenodo.org/record/3968687/files/LBBG_ZEEBRUGGE-gps-2013.csv.zip?download=1}. More information on Zenodo: https://zenodo.org/record/3968687.
}
\usage{
lbbg_gps
}
\description{
A bird tracking dataset containing 100 rows of GPS data from 2013 of Lesser black-backed gulls.
}
\keyword{datasets}
