% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbbg_reference}
\alias{lbbg_reference}
\title{Movebank reference data of a Lesser black-backed gulls bird tracking dataset}
\format{
A tibble (data.frame) with 100 rows and 24 variables.
}
\source{
Reference data in csv format: \url{https://zenodo.org/record/3968687/files/LBBG_ZEEBRUGGE-gps-2013.csv.zip?download=1}. More information on Zenodo: https://zenodo.org/record/3968687.
}
\usage{
lbbg_reference
}
\description{
A dataset containing reference data (metadata) of a GPS data from 2013 of lesser black-backed gulls.
}
\keyword{datasets}
