% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{load_bird_file}
\alias{load_bird_file}
\title{Load bird metadata file.}
\usage{
load_bird_file(filename)
}
\arguments{
\item{filename}{The name of the file containing bird metadata.}
}
\value{
A data table (not a data frame!) containing the bird metadata.
}
\description{
Load a file containing bird metadata. This file is managed at the INBO on
Google Drive. Create a csv export of that file, and make sure it is \code{,}
delimited.
}
\examples{
\dontrun{
load_bird_file(inputFile)
}
}
