% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{load_tracks_file}
\alias{load_tracks_file}
\title{Load file containing bird tracking data}
\usage{
load_tracks_file(filename)
}
\arguments{
\item{filename}{The file containing bird tracking data in csv
format. (','-delimited, header included)}
}
\value{
A data table (not a data frame!) containing the tracking data
}
\description{
Load a file containing bird tracking data. This file can
be obtained by requesting a dump from the UvA-BiTS virtual lab.
}
\examples{
\dontrun{
load_tracks_file(inputFile)
}
}
