% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profiling.R
\name{profiling_L1}
\alias{profiling_L1}
\title{Profile UvA-BiTS ETL processing steps (Level 1)}
\usage{
profiling_L1(tracks_file, birds_file)
}
\arguments{
\item{tracks_file}{CSV file containing tracking data used for profiling.
Don't use a file that is too big (because it will take some time), but don't
make it too small either (because that can introduces biases that are not
representative)}

\item{birds_file}{CSV file containing bird tracking metadata.}
}
\value{
Data.Frame with two columns: "step" and "timing"
}
\description{
Profile the different processing steps in the UvaBitsETL
package to find bottlenecks.
}

