% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{raster_join}
\alias{raster_join}
\title{Raster join}
\usage{
raster_join(dt, raster_data)
}
\arguments{
\item{dt}{Data table with tracking data. Geospatial points are created based on the
latitude and longitude field and WGS84 datum is expected.}

\item{raster_data}{A raster data object. See the package "raster".}
}
\value{
nothing. The value of the raster for every point is added to the data
table in place
}
\description{
Join tracks with raster data
}
