% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{read_raster_data}
\alias{read_raster_data}
\title{Read raster data.}
\usage{
read_raster_data(filename, data.CRS = "+init=epsg:4326")
}
\arguments{
\item{filename}{Name of the file containing raster data.}

\item{data.CRS}{Coordinate Reference System of the data.}
}
\value{
Raster data as RasterLayer class.
}
\description{
Read raster data using the raster package. By default this function will set
the CRS of this data to EPSG4326 (WGS 84). Use the data.CRS parameter to
override this.
}
\examples{
\dontrun{
read_raster_data("raster_file.tiff", data.CRS="+init=epsg:2056")
}
}
