% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sex_choices}
\alias{sex_choices}
\title{Sex choices}
\format{character vector}
\usage{
sex_choices
}
\description{
Allowed values for the sex column in bird metadata
}
\keyword{datasets}

