% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_choices}
\alias{species_choices}
\title{Species choices}
\format{character vector}
\usage{
species_choices
}
\description{
Allowed values for the species column in bird metadata
}
\keyword{datasets}
