% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{suncalc.custom}
\alias{suncalc.custom}
\title{Calculate sunrise and sunset.}
\usage{
suncalc.custom(dates, Lat, Long)
}
\arguments{
\item{dates}{Vector containing POSIX dates.}

\item{Lat}{Latitudes in WGS84.}

\item{Long}{Longitudes in WGS84.}
}
\value{
Named list with elements \code{sunrise} and \code{sunset}.
}
\description{
This function calculate the sunrise and sunset in UTC for a given location
and date. Function was slightly modified but inspired from this post:
\url{http://www.r-bloggers.com/approximate-sunrise-and-sunset-times/}
}
\examples{
{
suncalc.custom(ymd("2015-01-01"), Lat=50.821, Long=4.366)
}
}
