% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracks_validated}
\alias{tracks_validated}
\title{Tracks validated.}
\format{data table}
\usage{
tracks_validated
}
\description{
An example data table containing validated bird tracking data (only 100
records). This is what you can expect after successfully loading and
validating tracking data.
}
\keyword{datasets}
