\name{UvaBitsETL-package}
\alias{UvaBitsETL-package}
\alias{UvaBitsETL}
\docType{package}
\title{
\packageTitle{UvaBitsETL}
}
\description{
\packageDescription{UvaBitsETL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UvaBitsETL}
\packageIndices{UvaBitsETL}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{UvaBitsETL}

Maintainer: \packageMaintainer{UvaBitsETL}
}
\references{
http://www.lifewatch.be
http://lifewatch.inbo.be
http://www.uva-bits.nl
}
