\name{UvaBitsWarehouse-package}
\alias{UvaBitsWarehouse-package}
\alias{UvaBitsWarehouse}
\docType{package}
\title{
\packageTitle{UvaBitsWarehouse}
}
\description{
\packageDescription{UvaBitsWarehouse}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UvaBitsWarehouse}
\packageIndices{UvaBitsWarehouse}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{UvaBitsWarehouse}

Maintainer: \packageMaintainer{UvaBitsWarehouse}
}
\references{
http://www.lifewatch.be
http://lifewatch.inbo.be
http://www.uva-bits.nl
}
