% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{validate_bird_data}
\alias{validate_bird_data}
\title{Validate bird data}
\usage{
validate_bird_data(bird_data)
}
\arguments{
\item{bird_data}{The bird metadata as a data table}
}
\value{
validated bird metadata as a data table if no errors are found.
}
\description{
Validate the bird metadata
}

