% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{validate_tracks_data}
\alias{validate_tracks_data}
\title{Validate tracking data}
\usage{
validate_tracks_data(tracks_data)
}
\arguments{
\item{tracks_data}{The tracking data as a data table}
}
\value{
validated tracking data as a data table if no errors are found.
}
\description{
Validate the data coming either from a csv file or
from the UvA-BiTS virtual lab directly.
}
\examples{
\dontrun{
validate_tracks_data(tracking_data)
}
}
