% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_dist_to_colony}
\alias{add_dist_to_colony}
\title{Add distance to colony}
\usage{
add_dist_to_colony(dt)
}
\arguments{
\item{dt}{tracking data as data.table. Should contains columns `latitude`,
`longitude`, `colony_latitude`, `colony_longitude`}
}
\value{
nothing. Adds distance (in meters) in place as columns `distance_to_colony`
}
\description{
calculates the distance from the GPS position to the colony
}
\examples{
\dontrun{
add_dist_to_colony(tracking_data)
}
}

