% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_dist_travelled}
\alias{add_dist_travelled}
\title{Add Distance travelled}
\usage{
add_dist_travelled(dt)
}
\arguments{
\item{dt}{tracking data as data.table}
}
\value{
nothing. Distance (in meters) is added in place (`calc_distance_diff`)
}
\description{
will calculate the distance travelled since previous GPS fix
}
\examples{
\dontrun{
add_dist_travelled(tracking_data)
}
}
