% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_dist_travelled}
\alias{add_dist_travelled}
\title{Add Distance travelled}
\usage{
add_dist_travelled(dt)
}
\arguments{
\item{tracking}{data as data.table}
}
\value{
nothing. Data is added in place
}
\description{
will calculate the distance travelled since previous GPS fix
}
\examples{
\dontrun{
add_dist_travelled(tracking_data)
}
}

