% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_speed}
\alias{add_speed}
\title{Add speed}
\usage{
add_speed(dt)
}
\arguments{
\item{dt}{tracking data as data.table. Should contain column `calc_distance_diff`
and column `calc_time_diff`}
}
\value{
nothing. Data is added in place as column `calc_speed_2d`
}
\description{
calculates the average 2 dimensional speed of the individual since the
previous GPS fix
}
\examples{
\dontrun{
add_speed(tracking_data)
}
}

