% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_sunlight}
\alias{add_sunlight}
\title{Presence of sunlight.}
\usage{
add_sunlight(dt)
}
\arguments{
\item{dt}{Tracking data as data.table. Should contain columns \code{latitude},
\code{longitude} and \code{date_time}.}
}
\value{
Nothing. Column \code{calc_sunlight} is added in place. This is a logical
vector, indicating whether sunlight was present at time and location of the
GPS fix.
}
\description{
Calculate the presence of sunlight for every GPS fix. This is done using the
\code{suncalc()} function of the package \code{RAtmosphere}. If the \code{date_time} of the
GPS fix is after sunrise and before sunset, presence of sunlight is \code{1}.
Otherwise it is set to \code{0}.
}
\examples{
\dontrun{
add_sunlight(tracking_data)
}
}
