% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_metadata.r
\name{append_metadata}
\alias{append_metadata}
\title{Merge Movebank GPS event data and reference data}
\usage{
append_metadata(
  gps,
  reference_data,
  reference_cols = c("animal-taxon", "tag-id", "animal-id", "animal-comments",
    "animal-life-stage", "animal-mass", "animal-sex", "deployment-comments")
)
}
\arguments{
\item{gps}{Movebank GPS data of class matrix, data.frame or data.table.  \code{gps} must contain at least the following columns:
\itemize{
\item \code{individual-taxon-canonical-name}
\item \code{tag-local-identifier}
\item \code{individual-local-identifier}
}}

\item{reference_data}{Movebank reference data of class matrix, data.frame or data.table. \code{reference_data} must contain at least the columns defined in \code{reference_cols}.}

\item{reference_cols}{Character. Vector with the names of the columns of \code{reference_data} to add to \code{gps}. \code{reference-cols} must contain at least the following columns:
\itemize{
\item \code{animal-taxon}
\item \code{tag-id}
\item \code{animal-id}
}

Default:
\code{c("animal-taxon", "tag-id","animal-id", "animal-comments", "animal-life-stage", "animal-mass", "animal-sex", "deployment-comments")}}
}
\value{
a data.table
}
\description{
This function adds specific Movebank reference data to GPS event data in a given order. GPS and reference data can be downloaded directly from \href{https://www.movebank.org}{movebank}. Examples: \href{https://www.movebank.org/panel_embedded_movebank_webapp?gwt_fragment=page=studies,path=study985143423}{Lesser Black-backed Gulls} and \href{https://www.movebank.org/panel_embedded_movebank_webapp?gwt_fragment=page=studies,path=study986040562}{Herring Gulls}
}
\examples{
\dontrun{
}
}
