% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_metadata.R
\name{append_metadata}
\alias{append_metadata}
\title{Append Movebank reference data to Movebank GPS data}
\usage{
append_metadata(gps, ref_data, ref_cols = NULL)
}
\arguments{
\item{gps}{data.frame, data.table or matrix. Movebank GPS data with at least
the columns used for the join: \code{individual-taxon-canonical-name},
\code{tag-local-identifier} and \code{individual-local-identifier}.}

\item{ref_data}{data.frame, data.table or matrix. Movebank reference data
with at least the columns defined in \code{ref_cols}.}

\item{ref_cols}{character vector. Column names of \code{ref_data} to be added to
\code{gps}. Defaults to \code{NULL}, adding all columns. When specified, it must at
least contain the columns used for the join: \code{tag-id}, \code{animal-id} and
\code{animal-taxon}. Columns that have same name as the columns of \code{gps} are
dropped with a warning.}
}
\value{
A data.table with the GPS data (all columns except those used in the
join) appended with the reference data (all columns defined in
\code{ref_cols}).
}
\description{
This function joins Movebank GPS data with Movebank reference (meta)data, on
the shared columns:
\itemize{
\item \code{gps.tag-local-identifier} = \code{ref_data.tag-id}
\item \code{gps.individual-local-identifier} = \code{ref_data.animal-id}
\item \code{gps.individual-taxon-canonical-name} = \code{ref_data.animal-taxon}
}
}
\details{
GPS and reference data can be downloaded from
\href{https://www.movebank.org}{Movebank}.
}
\examples{
# Default use
append_metadata(lbbg_gps, lbbg_ref_data)

# Only include specific reference data columns
append_metadata(lbbg_gps,
                lbbg_ref_data,
                ref_cols = c("animal-taxon",
                             "tag-id",
                             "animal-id",
                             "animal-comments",
                             "animal-life-stage",
                             "animal-mass",
                             "animal-sex",
                             "deployment-comments")
)
}
