% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrich.R
\name{flag_outliers}
\alias{flag_outliers}
\title{Flag outliers}
\usage{
flag_outliers(dt)
}
\arguments{
\item{dt}{tracking data as data.table.}
}
\value{
nothing. Flagging happens in place. New columns is called `outlier`
and contains logical values.
}
\description{
Flag records that are suspect to be erronous.
The following checks are made:
    - date_time < current date
    - altitude < 1000 km
    - speed < 33.3333 meters per second (= 120 km per hour)
    - height_accuracy < 1000
If one of these fails, the record gets flagged.
}
\examples{
\dontrun{
flag_outliers(tracking_data)
}
}

