% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbbg_reference}
\alias{lbbg_reference}
\title{Sample of gull reference data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 156 rows and 24 columns.
}
\source{
\href{https://doi.org/10.5281/zenodo.3968687}{Stienen et al. 2020}
}
\usage{
lbbg_reference
}
\description{
A sample dataset of gull reference data in Movebank format.
}
\examples{
\dontrun{
ref_data <- read_csv("https://zenodo.org/record/3968687/files/LBBG_ZEEBRUGGE-reference-data.csv")
save(ref_data, file = "data/lbbg_reference.rda")
}
}
\keyword{datasets}
